v2.8.0 Bug Fix, Performance, Toast Notifications - Goal.Id auto-increment, Deadline DateOnly, Password policy, Query optimization, Database indexes

+++ FamilyFinance/Services/Interfaces/IGoalService.cs
@@ -5,8 +5,8 @@ namespace FamilyFinance.Services.Interfaces;
 public interface IGoalService
 {
     Task<List<Goal>> GetAllAsync(int familyId);
-    Task<Goal?> GetByIdAsync(long id);
+    Task<Goal?> GetByIdAsync(int id);
     Task SaveAsync(Goal goal);
-    Task DeleteAsync(long id);
+    Task DeleteAsync(int id);
 }
